Moving ECS files to sub module
Refactored base module's ECS files in to ecs module.
