+++ ecs/src/main/java/com/avogine/ecs/EntityArchetype.java
