+++ ecs/src/main/java/com/avogine/ecs/EntityComponent.java
