+++ ecs/src/test/java/com/avogine/ecs/components/TestWonkyArchetype.java
